kopano-webapp-lighttpd: enable additional package
Reworking the existing configuration file for Lighttpd and remove all
entries that are not useful for a configuration. In the end it contains
only a URL Alias to make kopano-webapp accessible as a subfolder on the
webserver if the user wants to use this.
The configuration isn't enabled by default (like the Apache2
configuration too) due the needed SSL configuration for safety.
Nevertheless it's usable also without a running SSL webserver config but
this needs to be enabled by the user.

Fixing the postinst script a bit more so it isn't failing while dpkg
runs.