autopkg: rename all references of zarafa into kopano
At time of writing the autopkg tests the depending packages are namend
'zarafa-*'. Zarafe did a re-naming of the old zarafa packages with
the new brand name 'kopano' an we need to adjust all old references to
packages or names of 'zarafa' to make the autopkgtests work again.

The old package kopano-webapp was split into separate packages for every
kind of webserver. Let's use the kopano-webapp-apache2 packge for
checking the functionality.