kopano-webapp: switch to upstream apparmor profile
Upstream has accepted a pull request to integrate the apparmor profile
which Guido has created and we already have integrated in the
kopano-webapp-apache2 package since 3.4.0+dfsg1-1.
For installation of this file from upstream we just need to adjust the
existing entry in the install sequencer file.
Thanks to Kopano for accepting the pull request about the apparmor
profile!.