use PHP_VERSION from one central place
We typically forget to increase the used PHP version in the various
files if we don't ude one central place to keep the ctual used version.
The file debian/rules is obviously *the* file we want to only change
once we need to bump or modify the PHP version. So set up a varable
PHP_VERSION here and use the value in the different places.

For example we use just sed to fill in the needed PHP version into the
postints script of kopano-webapp-nginx.