k-w-lighttpd: update README about http usage
Unfortunately we still can't provide a useful https configuration for
the lighttpd based packaging. As the prepared configuration in
kopano-webapp-lighttpd is based on a really simply http alias config,
but Kopano also added some check so people couldn't run a http setup
without additional needed modification to the setup, we added some extra
information to the README of the package. Users of this package will
need to look at the README anyhow as we don't enable the configuration
for lighttpd while package installation.